+++ src/main/java/edu/illinois/library/cantaloupe/operation/overlay/StringOverlay.java
@@ -123,7 +123,7 @@ public class StringOverlay extends Overlay implements Operation {
      */
     public void setString(String string) {
         checkFrozen();
-        this.string = string;
+        this.string = string.replace("\\n", "\n");
     }
 
     /**